common css file for both forms
