+++ src/components/Auth/Auth.jsx
@@ -1,9 +1,10 @@
 import { Link, useParams } from "react-router-dom";
-import RegisterForm from "../RegisterForm/RegisterForm";
-import LoginForm from "../LoginForm/LoginForm";
+
 import { useState } from "react";
 // import clsx from "clsx";
 import css from "./Auth.module.css";
+import RegisterForm from "../Form/RegisterForm";
+import LoginForm from "../Form/LoginForm";
 const Auth = () => {
   const { id } = useParams();
   const [register, setRegister] = useState(id === "register");
