+++ src/components/Form/Form.css
@@ -1,4 +1,4 @@
-/* .form {
+.form {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -53,4 +53,4 @@
   transform: translateY(-50%);
   cursor: pointer;
   color:  rgba(255, 255, 255, 0.4); 
-} */
+}
