+++ src/components/LoginForm/LoginForm.module.css
@@ -1,4 +1,4 @@
-.form {
+/* .form {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -27,4 +27,4 @@
   margin-bottom: 15px;
   font-size: 16px;
   outline: #1F1F1F;
-}
+} */
